refactor: merge `SlideItem` and `SlidesCarousel` into one `Slides` component
